Implement an inline assembly fallback for unsupported instructions. (#232)
This won't support all cases (256-bit and 512-bit instructions for sure), but
this should allow for correctness on unknown instructions, even if the generated
output will cause the optimizers to throw their hands up in the air.

The test is somewhat flaky since it passes 16 arguments into printf after the
the printf string, which seems to be more than mcsema expects. It also requires
particular compilation flags to get the aes instructions to compile and run
(that means you need Intel Westmere or newer or AMD Bulldozer or newer).

I used AES for the test case since I figured no one would ever bother
implementing an instruction lifter for these instructions unless it were part
of a generalized intrinsic uplifting technique.