Immediate values, partial variables
Rework code around immediate values, it is finally not just blindly
adding xrefs from everything

Sometimes two variables can occupy the same space (if one is suffix of
other). There is no real way to resolve it, so if there is something
targeting .data, .rodata, .bss it is just added as xref.