Docs: Update the README to include libmagic; use pth instead of PYTHONPATH
The Binary Ninja scripts import the magic module, causing a failure if the
native libmagic.dll library is not found. This patch updates the README
to include the correct pip install command.

This patch also removes the suggestion to use PYTHONPATH, since it then
hides the local site-packages folder. The new recommended way is to
create a .pth file in the site-packages that includes the mcsema
modules.