Windows 10/Visual Studio 2019 compatibility (#593)
* Docs: Update the README for Visual Studio 2019

* CMake: Fix the Python module install step on Windows

* Docs: Update the step numbers in the README

* Docs: Do not hardcode the build job count in the README

* Docs: Update the README to document how to setup the environment

* Docs: Update the README to include libmagic; use pth instead of PYTHONPATH

The Binary Ninja scripts import the magic module, causing a failure if the
native libmagic.dll library is not found. This patch updates the README
to include the correct pip install command.

This patch also removes the suggestion to use PYTHONPATH, since it then
hides the local site-packages folder. The new recommended way is to
create a .pth file in the site-packages that includes the mcsema
modules.

* CMake: Fix regression introduced in the install step for Linux

* Update .remill_commit_id